analytics-svc/requirements.txt
APScheduler==3.11.0
dnspython==2.7.0
mysql-connector-python==9.2.0
pymongo==4.11.1
tzlocal==5.3
